+++ Plugins/Myra/Source/Myra/Public/Character/MyraPawnExtensionComponent.h
@@ -12,7 +12,6 @@ class UEnhancedInputComponent;
 class UInputComponent;
 class UMyraAbilitySystemComponent;
 class UMyraPawnData;
-class AMyraCharacter;
 
 DECLARE_DYNAMIC_MULTICAST_DELEGATE(FMyraOnPawnReadyToInitialize);